Reverse compound interest formula in JAVA
It's not a template, but rather an example. Used this code to answer wrongly asked question in my 3rd semester at Humber College in 2019 on Kristian Medri's 2nd home/lab assignment on Java course on Computer Engineering Program. I knew that I could have answered the original question the easy (expected) way, but I decided to answer it word to word (the way it was interpreted) and create a formula that (at least back then) did not exist on the internet.
I have uploaded this code on StackOverflow back in autumn 2020 and almost a year later (today - 17th Aug 2021), BUT BOTH TIMES IT GOT CANCELLED, HIDDEN, AND DELETED due to "lack of clarity"