Excel - Contact list & form example
Based on the https://youtu.be/SnA3cct-2IM tutorial. 
Used VBA. Fixed one bug from the video: saving new contact caused a 1004 error. It was because the B2 was empty when it was supposed to have a value.
Made this excel at the supervisor's request. Had to go a long way from thinking that excel can't have forms, then making it on Access, then realizing that Access doesn't run on Mac, then making it on LibreOffice Base, then getting frustrated with LO Base, realizing that I've never googled "excel FORM", then going back to excel to make the final form+list.